webpack/refactor: Prepare JS files for consumption by webpack..
This commits prepares the frontend code to be consumed by webpack.
In order to do this we expose the local variable name to the global
window context by setting window.varName = varName; as needed in
the js files.

It also edits the front-end node tests to provide a default
window object which is assumed to exist by the frontend code.
To do this we set a global window object and use the the to_$
hack to enable it being wrapped by $(window) calls in the test code.

It refactors  frontend tests to use the patch_builtin function
instead of the set_global function for window so it is restored
to its default state properly when the test has finished running.

Since app-styles will now be included by app.js we need to change
the backend test accordingly.