webpack: Run multiple bundles on same page efficiently.
This commit enables us to run multiple webpack bundles on the
same page efficiently.

This is achieved by telling webpack to write the runtime to a
separate chunk called runtime. In order to use these in Django
we use 3 plugins for webpack to be able to include the runtime
in the Django template ( since the runtime is different each
time a different version of the app is built ).

1) html-webpack-plugin
2) inline-manifest-webpack-plugin
3) html-webpack-harddisk-plugin

Each do their job to be able extract it, inline it to a html
template and write it to disk. Subsequently that dynamically
generated webpack_runtime_template is then 'included'
by base.html template in Django.

This enables multiple modules to run on the same
page which would otherwise cause issues like expose-loader
not applying across different bundles, multiple versions
of the runtime on the same page etc.

Potentially this hack could be replaced by using the built-in
webpack code-splitting feature however, we're currently not
able to use it easily due to django-webpack-loader plugin
not supporting this feature for Webpack 4.

See https://github.com/owais/django-webpack-loader/issues/157
and https://github.com/owais/django-webpack-loader/issues/165